requirements.txt
flask==3.1.0
flask-session
redis==5.2.1
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.33
langchain-ollama==0.2.3
langchain-text-splitters==0.3.5
faiss-cpu
pymupdf
python-dotenv==1.0.1
rank_bm25